refactor dataset wrapper to prepare dali wrapper
